+++ src/azure-cli/azure/cli/command_modules/acs/_params.py
@@ -323,6 +323,8 @@ def load_arguments(self, _):
         c.argument('customer_admin_group_id', options_list=['--customer-admin-group-id'])
         c.argument('workspace_resource_id')
 
+    with self.argument_context('openshift monitor enable') as c:
+        c.argument('workspace-resource-id')
 
 def _get_default_install_location(exe_name):
     system = platform.system()